% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{fun}
\alias{fun}
\title{A simply incredible function}
\usage{
fun(x)
}
\arguments{
\item{x}{Some description here.}
}
\value{
We return the x value back!
}
\description{
A simply incredible function
}
\examples{
x <- 'an amazing character vector'
fun(x)
}
