% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{is.dsmm}
\alias{is.dsmm}
\title{Function to check if an object is of class \code{`dsmm`}}
\usage{
is.dsmm(obj)
}
\arguments{
\item{obj}{Arbitrary \code{`R`} object.}
}
\value{
`TRUE` or `FALSE`.
}
\description{
It checks for inheritance of the class \code{`dsmm`} and
 also for the validity of the attributes specified.
}
\seealso{
\link{is.dsmm_fit}, \link{is.dsmm_parametric},
\link{is.dsmm_nonparametric}
}
