% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{is.dsmm}
\alias{is.dsmm}
\title{Check if an object has a valid \code{dsmm} class}
\usage{
is.dsmm(obj)
}
\arguments{
\item{obj}{Arbitrary \code{R} object.}
}
\value{
TRUE or FALSE.
}
\description{
Checks for the validity of the specified attributes  and the
   inheritance of the S3 class \code{dsmm}. This class acts like a parent
   class for the classes \code{dsmm_fit_nonparametric,}
   \code{dsmm_fit_parametric, dsmm_parametric} and \code{dsmm_nonparametric}.
}
\seealso{
\link{is.dsmm_fit_nonparametric}, \link{is.dsmm_fit_nonparametric},
     \link{is.dsmm_parametric}, \link{is.dsmm_nonparametric}
}
