% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{is.dsmm_fit_nonparametric}
\alias{is.dsmm_fit_nonparametric}
\title{Check if an object has a valid \code{dsmm_fit_nonparametric} class}
\usage{
is.dsmm_fit_nonparametric(obj)
}
\arguments{
\item{obj}{Arbitrary \code{R} object.}
}
\value{
TRUE or FALSE.
}
\description{
Checks for the validity of the specified attributes and the
    inheritance of the S3 class \code{dsmm_fit_nonparametric}.
    This class inherits methods from the parent class \code{dsmm}.
}
\seealso{
\link{is.dsmm}, \link{is.dsmm_fit_parametric},
     \link{is.dsmm_nonparametric}, \link{is.dsmm_parametric}
}
